Skip check of initial indices if they have been rolled-over already
We are getting list of initial indices without write alias to operate
on later (i.e. we create write alias for them).
There can be, however, a valid case when initial index had the write
alias but after the rollover a new index has been created and the write
alias was pointed at this new index.

The situation looks like this:

app-foo-000001
app-foo-000002 <- app-foo-write

In this case we need to skip the initial index without write alias.
It is valid case and we shall not try to create write alias for it.
This was causing unnecessary exception being thrown out.